+++ src/ffi/rust.rs
@@ -41,7 +41,7 @@ impl fmt::Debug for Inflate {
 }
 
 impl InflateBackend for Inflate {
-    fn make(zlib_header: bool, window_bits: u8) -> Self {
+    fn make(zlib_header: bool, window_bits: i32) -> Self {
         assert!(
             window_bits > 8 && window_bits < 16,
             "window_bits must be within 9 ..= 15"
@@ -119,7 +119,7 @@ impl fmt::Debug for Deflate {
 }
 
 impl DeflateBackend for Deflate {
-    fn make(level: Compression, zlib_header: bool, window_bits: u8) -> Self {
+    fn make(level: Compression, zlib_header: bool, window_bits: i32) -> Self {
         assert!(
             window_bits > 8 && window_bits < 16,
             "window_bits must be within 9 ..= 15"
